+++ src/utils/searchParameterConstructor.js
